refactor(anguar): migrate router link directive to common
